requirements.txt
streamlit
psycopg2
pandas
requests
python-dotenv
sqlalchemy
